Add billable_locations route

+++ db/migrate/20190924141217_create_people.rb
@@ -2,7 +2,7 @@ class CreatePeople < ActiveRecord::Migration[6.0]
   def change
     create_table :people do |t|
       t.string      :name
-      t.decimal     :salary
+      t.integer     :salary
       t.belongs_to  :location
       t.belongs_to  :role
       t.references  :manager, null: true
